requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
httpx>=0.28.1
numpy>=1.24.0
pandas>=2.0.0
python-dateutil>=2.8.2
aiohttp>=3.8.4
asyncio>=3.4.3
solders>=0.19.0
solana>=0.30.0
websockets>=10.3
pylint>=2.15.0
jupyterlab>=3.5.0
matplotlib>=3.6.0
python-telegram-bot>=22.5
